Set up classes and seed database
This commit sets up the Ingredient and Recipe classes as well as the recipes_controller. It also seeds the database, which is structured as a recipes table, an ingredients table and a recipe_ingredients join table.

Users can now view a list of all recipes at the `/recipes` endpoint. They can also click the titles of those recipes to be taken to the full recipe.