WILL REMOVE - re-adding ingredients to recipes
This was the beginning of an ill-advised foray into adding an array of ingredient ids onto the recipes table. I was trying a few things out and thought I might want to venture down this path but have since changed my mind, so will need another migration to undo this one (I have an existing migration that already does this so can just copy that). But it's now late :)